Do not send update diagram on interactive movement.
