+++ applications/klighd-vscode/package.json
@@ -161,6 +161,7 @@
     },
     "dependencies": {
         "@kieler/klighd-core": "^0.4.0",
+        "@kieler/klighd-interactive": "^0.4.0",
         "inversify": "^5.0.1",
         "nanoid": "^3.1.23",
         "reflect-metadata": "^0.1.13",
